refactor: move web action into web-actions from bridge-actions
